+++ meetnote3/src/nativeMain/kotlin/meetnote3/utils/ProcessInfo.kt
@@ -1,0 +1,10 @@
+package meetnote3.utils
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ProcessInfo(
+    val pid: Int,
+    val ppid: Int,
+    val name: String,
+)
